+++ src/localization/include/localizer/localizer.h
@@ -11,9 +11,9 @@ class Localizer : public LocalizerInterface::LocalizerInterface_c
 {
 public:
   LocalizerInterface::stateVector getStateVector();
-  UpdateStatus updateStateVector(float dt);
+  UpdateStatus updateStateVector(double dt);
 
-  Localizer(float axelLen, float xi, float yi, float thi, iVescAccess *frontLeftVesc, iVescAccess *frontRightVesc,
+  Localizer(double axelLen, double xi, double yi, double thi, iVescAccess *frontLeftVesc, iVescAccess *frontRightVesc,
             iVescAccess *backRightVesc,
             iVescAccess *backLeftVesc);  // pass wheel linear vel sensors in as FL, FR, BR, BL
 
@@ -27,7 +27,7 @@ protected:
   iVescAccess *back_left_vesc;
   // int timediffms(struct timeval curr, struct timeval prev);
   LocalizerInterface::stateVector state_vector;
-  float axle_len;
+  double axle_len;
 };
 
 #endif
