+++ src/localization/include/super_localizer/super_localizer.h
@@ -14,13 +14,13 @@
 class SuperLocalizer : public LocalizerInterface::LocalizerInterface_c
 {
 public:
-  SuperLocalizer(float axleLen, float xi, float yi, float thi, iVescAccess *frontLeftVesc, iVescAccess *frontRightVesc,
+  SuperLocalizer(double axleLen, double xi, double yi, double thi, iVescAccess *frontLeftVesc, iVescAccess *frontRightVesc,
                  iVescAccess *backRightVesc, iVescAccess *backLeftVesc, ImuSensorInterface *centerIMU,
                  PosSensorInterface *posSensor, LocalizerInterface::stateVector gains);
-  SuperLocalizer(float axleLen, float xi, float yi, float thi, iVescAccess *frontLeftVesc, iVescAccess *frontRightVesc,
+  SuperLocalizer(double axleLen, double xi, double yi, double thi, iVescAccess *frontLeftVesc, iVescAccess *frontRightVesc,
                  iVescAccess *backRightVesc, iVescAccess *backLeftVesc, PosSensorInterface *posSensor,
                  LocalizerInterface::stateVector gains);
-  UpdateStatus updateStateVector(float dt);
+  UpdateStatus updateStateVector(double dt);
   ~SuperLocalizer();
 
   // static constexpr Localizer::stateVector_s default_gains;
@@ -47,7 +47,7 @@ private:
 
   LocalizerInterface::stateVector state_vector;
 
-  LocalizerInterface::stateVector initState(float xi, float yi, float theta);
+  LocalizerInterface::stateVector initState(double xi, double yi, double theta);
 };
 
 // I couldn't figure out how to make this a static class member. I tried quite a few things...
