+++ src/localization/src/localizer/localizer.cpp
@@ -3,7 +3,7 @@
 #include <cmath>
 #define CONSTANT_TO_AVERAGE_TWO_NUMBERS 2.0f
 
-Localizer::Localizer(float axleLen, float xi, float yi, float thi, iVescAccess *frontLeftVesc,
+Localizer::Localizer(double axleLen, double xi, double yi, double thi, iVescAccess *frontLeftVesc,
                      iVescAccess *frontRightVesc, iVescAccess *backRightVesc, iVescAccess *backLeftVesc)
 {
   state_vector.x_pos = xi;
@@ -29,23 +29,23 @@ Localizer::Localizer(float axleLen, float xi, float yi, float thi, iVescAccess *
   axle_len = axleLen;
 }
 
-Localizer::UpdateStatus Localizer::updateStateVector(float dt)
+Localizer::UpdateStatus Localizer::updateStateVector(double dt)
 {
   // get linear velocities of wheels
-  float front_left_velocity = this->front_left_vesc->getLinearVelocity();
-  float front_right_velocity = this->front_right_vesc->getLinearVelocity();
-  float back_right_velocity = this->back_right_vesc->getLinearVelocity();
-  float back_left_velocity = this->back_left_vesc->getLinearVelocity();
+  double front_left_velocity = this->front_left_vesc->getLinearVelocity();
+  double front_right_velocity = this->front_right_vesc->getLinearVelocity();
+  double back_right_velocity = this->back_right_vesc->getLinearVelocity();
+  double back_left_velocity = this->back_left_vesc->getLinearVelocity();
 
-  float average_left_velocity = (front_left_velocity + back_left_velocity) / CONSTANT_TO_AVERAGE_TWO_NUMBERS;
-  float average_right_velocity = (front_right_velocity + back_right_velocity) / CONSTANT_TO_AVERAGE_TWO_NUMBERS;
+  double average_left_velocity = (front_left_velocity + back_left_velocity) / CONSTANT_TO_AVERAGE_TWO_NUMBERS;
+  double average_right_velocity = (front_right_velocity + back_right_velocity) / CONSTANT_TO_AVERAGE_TWO_NUMBERS;
 
-  float w = (average_right_velocity - average_left_velocity) / axle_len;
-  float turn_radius;
+  double w = (average_right_velocity - average_left_velocity) / axle_len;
+  double turn_radius;
   Eigen::Matrix2f rot;
   Eigen::Vector2f d_pos;
   Eigen::Vector2f rotation_on_y;
-  float d_theta;
+  double d_theta;
 
   if (std::abs(average_right_velocity - average_left_velocity) > 0.01f)  // no dividing by zero
   {
